Add selector for YouTube Up Next cancel button
In some views such as the "miniplayer" view, there is no autoplay switch
available. In this case we keep watching the DOM and click the "cancel" button
whenever the Up Next card appears at the end of a video.

This may also serve as a fallback in future if the current selectors for the
autoplay switch stop working.